pyproject.toml
[project]
name = "sempervigil"
version = "0.1.0"
description = "SemperVigil ingestion pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0" }
authors = [{ name = "SemperVigil" }]
dependencies = [
  "feedparser>=6.0.11",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
test = ["pytest>=7.4"]

[project.scripts]
sempervigil = "sempervigil.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
